Automatically backed up by Learn

+++ lib/bachelor.rb
@@ -60,7 +60,7 @@ count = 0
     people.each do |k,v|
       if k == "age"
         count += 1
-        age += v.to_f
+        age += v.to_i
       end
     end
   end
